+++ ios/platform/GoogleMobileAds.framework/Headers/GADMobileAds.h
@@ -20,6 +20,9 @@ typedef void (^GADInitializationCompletionHandler)(GADInitializationStatus *_Non
 /// Returns the shared GADMobileAds instance.
 + (nonnull GADMobileAds *)sharedInstance;
 
+/// Returns the version of the SDK.
+@property(nonatomic, nonnull, readonly) NSString *sdkVersion;
+
 /// The application's audio volume. Affects audio volumes of all ads relative to other audio output.
 /// Valid ad volume values range from 0.0 (silent) to 1.0 (current device volume). Use this method
 /// only if your application has its own volume controls (e.g., custom music or sound effect