+++ ios/platform/GoogleMobileAds.framework/Headers/GADNativeAd.h
@@ -8,6 +8,9 @@
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
 
+#import <GoogleMobileAds/GADResponseInfo.h>
+#import <GoogleMobileAds/GoogleMobileAdsDefines.h>
+
 @protocol GADNativeAdDelegate;
 
 /// Native ad base class. All native ad types are subclasses of this class.
@@ -24,9 +27,14 @@
 /// Dictionary of assets which aren't processed by the receiver.
 @property(nonatomic, readonly, copy, nullable) NSDictionary *extraAssets;
 
+/// Information about the ad response that returned the ad.
+@property(nonatomic, readonly, nonnull) GADResponseInfo *responseInfo;
+
 /// The ad network class name that fetched the current ad. For both standard and mediated Google
 /// AdMob ads, this method returns @"GADMAdapterGoogleAdMobAds". For ads fetched via mediation
 /// custom events, this method returns @"GADMAdapterCustomEvents".
-@property(nonatomic, readonly, copy, nullable) NSString *adNetworkClassName;
+@property(nonatomic, readonly, copy, nullable)
+    NSString *adNetworkClassName GAD_DEPRECATED_MSG_ATTRIBUTE(
+        "Use responseInfo.adNetworkClassName.");
 
 @end