+++ ios/documentation/changelog.md
@@ -1,4 +1,8 @@
 # Change Log
+v2.5.2  Update AdMob SDK to 7.64.0
+        Update Audience Network SDK 5.10.1
+        Update Facebook Adapter 5.10.1.0
+
 <pre>
 v2.4.4  Update AdMob SDK to 7.59.0
         Update PersonalizedAdConsent to 1.5.0 (Removed all references to deprecated UIWebView)